+++ CideFlutter/lib/providers/ide_notifier.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:cide/src/rust/api/types.dart' as rust;
 import 'package:cide/src/rust/compiler/intent.dart';
+import '../services/rust_api_service.dart';
 import '../models/algorithm_validation.dart';
 import '../models/code_template.dart';
 import '../models/ide_state.dart';