+++ CideFlutter/lib/providers/ide_provider.dart
@@ -4,5 +4,6 @@ import 'ide_notifier.dart';
 
 export 'ide_notifier.dart';
 export '../models/ide_state.dart';
+export '../services/rust_api_service.dart' show RustApiService, rustApiServiceProvider;
 
 final ideProvider = AutoDisposeNotifierProvider<IdeNotifier, IdeState>(IdeNotifier.new);